se agregan nuevas clases y metodos para - comic - pelicula y se agregan los metodos borrar insertar actualizar para las demas clases que ya estaban hechas

+++ cl_marvel_superhero/src/main/java/cl/marvel/superhero/repo/ICaracteristicaRepo.java
@@ -3,6 +3,6 @@ package cl.marvel.superhero.repo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import cl.marvel.superhero.model.Caracteristica;
 
-public interface ICaracteristicaRepo extends JpaRepository<Caracteristica, Integer> {
+public interface ICaracteristicaRepo extends JpaRepository<Caracteristica, Long> {
 
 }
